+++ fp_ip_test/Ep3Code/qbRayTrace/camera.cpp
@@ -30,94 +30,94 @@
 qbRT::Camera::Camera()
 {
 	// The default constructor.
-	m_cameraPosition = qbVector<double>	{std::vector<double> {0.0, -10.0, 0.0}};
-	m_cameraLookAt = qbVector<double>		{std::vector<double> {0.0, 0.0, 0.0}};
-	m_cameraUp = qbVector<double>				{std::vector<double> {0.0, 0.0, 1.0}};
+	m_cameraPosition = qbVector<float>	{std::vector<float> {0.0, -10.0, 0.0}};
+	m_cameraLookAt = qbVector<float>		{std::vector<float> {0.0, 0.0, 0.0}};
+	m_cameraUp = qbVector<float>				{std::vector<float> {0.0, 0.0, 1.0}};
 	m_cameraLength = 1.0;
 	m_cameraHorzSize = 1.0;
 	m_cameraAspectRatio = 1.0;
 }
 
-void qbRT::Camera::SetPosition(const qbVector<double> &newPosition)
+void qbRT::Camera::SetPosition(const qbVector<float> &newPosition)
 {
 	m_cameraPosition = newPosition;
 }
 
-void qbRT::Camera::SetLookAt(const qbVector<double> &newLookAt)
+void qbRT::Camera::SetLookAt(const qbVector<float> &newLookAt)
 {
 	m_cameraLookAt = newLookAt;
 }
 
-void qbRT::Camera::SetUp(const qbVector<double> &upVector)
+void qbRT::Camera::SetUp(const qbVector<float> &upVector)
 {
 	m_cameraUp = upVector;
 }
 
-void qbRT::Camera::SetLength(double newLength)
+void qbRT::Camera::SetLength(float newLength)
 {
 	m_cameraLength = newLength;
 }
 
-void qbRT::Camera::SetHorzSize(double newHorzSize)
+void qbRT::Camera::SetHorzSize(float newHorzSize)
 {
 	m_cameraHorzSize = newHorzSize;
 }
 
-void qbRT::Camera::SetAspect(double newAspect)
+void qbRT::Camera::SetAspect(float newAspect)
 {
 	m_cameraAspectRatio = newAspect;
 }
 
 // Method to return the position of the camera.
-qbVector<double> qbRT::Camera::GetPosition()
+qbVector<float> qbRT::Camera::GetPosition()
 {
 	return m_cameraPosition;
 }
 
 // Method to return the LookAt of the camera.
-qbVector<double> qbRT::Camera::GetLookAt()
+qbVector<float> qbRT::Camera::GetLookAt()
 {
 	return m_cameraLookAt;
 }
 
 // Method to return the up vector of the camera.
-qbVector<double> qbRT::Camera::GetUp()
+qbVector<float> qbRT::Camera::GetUp()
 {
 	return m_cameraUp;
 }
 
 // Method to return the length of the camera.
-double qbRT::Camera::GetLength()
+float qbRT::Camera::GetLength()
 {
 	return m_cameraLength;
 }
 
 // Method to return the horizontal size.
-double qbRT::Camera::GetHorzSize()
+float qbRT::Camera::GetHorzSize()
 {
 	return m_cameraHorzSize;
 }
 
 // Method to return the camera aspect ratio.
-double qbRT::Camera::GetAspect()
+float qbRT::Camera::GetAspect()
 {
 	return m_cameraAspectRatio;
 }
 
 // Method to return the U vector.
-qbVector<double> qbRT::Camera::GetU()
+qbVector<float> qbRT::Camera::GetU()
 {
 	return m_projectionScreenU;
 }
 
 // Method to return the V vector.
-qbVector<double> qbRT::Camera::GetV()
+qbVector<float> qbRT::Camera::GetV()
 {
 	return m_projectionScreenV;
 }
 
 // Method to return the projection screen centre.
-qbVector<double> qbRT::Camera::GetScreenCentre()
+qbVector<float> qbRT::Camera::GetScreenCentre()
 {
 	return m_projectionScreenCentre;
 }
@@ -130,9 +130,9 @@ void qbRT::Camera::UpdateCameraGeometry()
 	m_alignmentVector.Normalize();
 	
 	// Second, compute the U and V vectors.
-	m_projectionScreenU = qbVector<double>::cross(m_alignmentVector, m_cameraUp);
+	m_projectionScreenU = qbVector<float>::cross(m_alignmentVector, m_cameraUp);
 	m_projectionScreenU.Normalize();
-	m_projectionScreenV = qbVector<double>::cross(m_projectionScreenU, m_alignmentVector);
+	m_projectionScreenV = qbVector<float>::cross(m_projectionScreenU, m_alignmentVector);
 	m_projectionScreenV.Normalize();
 	
 	// Thirdly, compute the positon of the centre point of the screen.
@@ -146,8 +146,8 @@ void qbRT::Camera::UpdateCameraGeometry()
 bool qbRT::Camera::GenerateRay(float proScreenX, float proScreenY, qbRT::Ray &cameraRay)
 {
 	// Compute the location of the screen point in world coordinates.
-	qbVector<double> screenWorldPart1 = m_projectionScreenCentre + (m_projectionScreenU * proScreenX);
-	qbVector<double> screenWorldCoordinate = screenWorldPart1 + (m_projectionScreenV * proScreenY);
+	qbVector<float> screenWorldPart1 = m_projectionScreenCentre + (m_projectionScreenU * proScreenX);
+	qbVector<float> screenWorldCoordinate = screenWorldPart1 + (m_projectionScreenV * proScreenY);
 	
 	// Use this point along with the camera position to compute the ray.
 	cameraRay.m_point1 = m_cameraPosition;
