+++ fp_ip_test/Ep3Code/qbRayTrace/camera.hpp
@@ -38,23 +38,23 @@ namespace qbRT
 			Camera();
 			
 			// Functions to set camera parameters.
-			void SetPosition	(const qbVector<double> &newPosition);
-			void SetLookAt		(const qbVector<double> &newLookAt);
-			void SetUp				(const qbVector<double> &upVector);
-			void SetLength		(double newLength);
-			void SetHorzSize	(double newSize);
-			void SetAspect		(double newAspect);
+			void SetPosition	(const qbVector<float> &newPosition);
+			void SetLookAt		(const qbVector<float> &newLookAt);
+			void SetUp				(const qbVector<float> &upVector);
+			void SetLength		(float newLength);
+			void SetHorzSize	(float newSize);
+			void SetAspect		(float newAspect);
 			
 			// Functions to return camera parameters.
-			qbVector<double>	GetPosition();
-			qbVector<double>	GetLookAt();
-			qbVector<double>	GetUp();
-			qbVector<double>	GetU();
-			qbVector<double>	GetV();
-			qbVector<double>	GetScreenCentre();
-			double						GetLength();
-			double						GetHorzSize();
-			double						GetAspect();
+			qbVector<float>	GetPosition();
+			qbVector<float>	GetLookAt();
+			qbVector<float>	GetUp();
+			qbVector<float>	GetU();
+			qbVector<float>	GetV();
+			qbVector<float>	GetScreenCentre();
+			float						GetLength();
+			float						GetHorzSize();
+			float						GetAspect();
 			
 			// Function to generate a ray.
 			bool GenerateRay(float proScreenX, float proScreenY, qbRT::Ray &cameraRay);
@@ -63,17 +63,17 @@ namespace qbRT
 			void UpdateCameraGeometry();
 			
 		private:
-			qbVector<double> m_cameraPosition	{3};
-			qbVector<double> m_cameraLookAt		{3};
-			qbVector<double> m_cameraUp				{3};
-			double m_cameraLength;
-			double m_cameraHorzSize;
-			double m_cameraAspectRatio;
+			qbVector<float> m_cameraPosition	{3};
+			qbVector<float> m_cameraLookAt		{3};
+			qbVector<float> m_cameraUp				{3};
+			float m_cameraLength;
+			float m_cameraHorzSize;
+			float m_cameraAspectRatio;
 			
-			qbVector<double> m_alignmentVector				{3};
-			qbVector<double> m_projectionScreenU			{3};
-			qbVector<double> m_projectionScreenV			{3};
-			qbVector<double> m_projectionScreenCentre	{3};
+			qbVector<float> m_alignmentVector				{3};
+			qbVector<float> m_projectionScreenU			{3};
+			qbVector<float> m_projectionScreenV			{3};
+			qbVector<float> m_projectionScreenCentre	{3};
 			
 	};
 }
