+++ fp_ip_test/Ep3Code/qbRayTrace/objectbase.cpp
@@ -41,13 +41,13 @@ qbRT::ObjectBase::~ObjectBase()
 }
 
 // Function to test for intersections.
-bool qbRT::ObjectBase::TestIntersection(const Ray &castRay, qbVector<double> &intPoint, qbVector<double> &localNormal, qbVector<double> &localColor)
+bool qbRT::ObjectBase::TestIntersection(const Ray &castRay, qbVector<float> &intPoint, qbVector<float> &localNormal, qbVector<float> &localColor)
 {
 	return false;
 }
 
 // Function to test whether two floating-point numbers are close to being equal.
-bool qbRT::ObjectBase::CloseEnough(const double f1, const double f2)
+bool qbRT::ObjectBase::CloseEnough(const float f1, const float f2)
 {
 	return fabs(f1-f2) < EPSILON;
 }
