+++ fp_ip_test/Ep3Code/qbRayTrace/ray.cpp
@@ -27,24 +27,24 @@
 
 qbRT::Ray::Ray()
 {
-	m_point1 = qbVector<double>{std::vector<double>{0.0, 0.0, 0.0}};
-	m_point2 = qbVector<double>{std::vector<double>{0.0, 0.0, 1.0}};
+	m_point1 = qbVector<float>{std::vector<float>{0.0, 0.0, 0.0}};
+	m_point2 = qbVector<float>{std::vector<float>{0.0, 0.0, 1.0}};
 	m_lab = m_point2 - m_point1;
 }
 
-qbRT::Ray::Ray(const qbVector<double> &point1, const qbVector<double> &point2)
+qbRT::Ray::Ray(const qbVector<float> &point1, const qbVector<float> &point2)
 {
 	m_point1 = point1;
 	m_point2 = point2;
 	m_lab = m_point2 - m_point1;
 }
 
-qbVector<double> qbRT::Ray::GetPoint1() const
+qbVector<float> qbRT::Ray::GetPoint1() const
 {
 	return m_point1;
 }
 
-qbVector<double> qbRT::Ray::GetPoint2() const
+qbVector<float> qbRT::Ray::GetPoint2() const
 {
 	return m_point2;
 }
