+++ fp_ip_test/Ep3Code/qbRayTrace/ray.hpp
@@ -34,15 +34,15 @@ namespace qbRT
 	{
 		public:
 			Ray();
-			Ray(const qbVector<double> &point1, const qbVector<double> &point2);
+			Ray(const qbVector<float> &point1, const qbVector<float> &point2);
 			
-			qbVector<double> GetPoint1() const;
-			qbVector<double> GetPoint2() const;
+			qbVector<float> GetPoint1() const;
+			qbVector<float> GetPoint2() const;
 			
 		public:
-			qbVector<double> m_point1	{3};
-			qbVector<double> m_point2 {3};
-			qbVector<double> m_lab		{3};
+			qbVector<float> m_point1	{3};
+			qbVector<float> m_point2 {3};
+			qbVector<float> m_lab		{3};
 			
 	};
 }
